+++ patches/src/main/kotlin/app/revanced/patches/spotify/misc/extension/Fingerprints.kt
@@ -3,5 +3,5 @@ package app.revanced.patches.spotify.misc.extension
 import app.revanced.patcher.fingerprint
 
 internal val loadOrbitLibraryFingerprint = fingerprint {
-    strings("OrbitLibraryLoader", "cst")
+    strings("orbit_library_load", "orbit-jni-spotify")
 }